+++ examples/polopt.jl
@@ -47,7 +47,7 @@ gmp = GMPModel()
  
 
 # Constrain μ to be a probablity measure
-@mconstraint gmp MomCon(Mom(1,μ), :eq, 1)
+@constraint gmp Mom(1,μ) == 1
 
 
 # We solve the relaxation of order 2 with CSDP
@@ -57,7 +57,7 @@ relax!(gmp, 2, with_optimizer(Mosek.Optimizer))
 println("Relaxation order: $(2)")
 println("Objective value: $(objective_value(gmp))")
 # We try to extract atoms from the relaxed moment sequence of μ
-opt = atomic(gmp,μ)
+opt = atomic(gmp, μ 1e-03)
 println()
 
 # As we could not extract atoms from the solution, we increase the relaxation order
@@ -65,7 +65,7 @@ relax!(gmp, 3, with_optimizer(Mosek.Optimizer))
 
 println("Relaxation order: $(3)")
 println("Objective value: $(objective_value(gmp))")
-opt = atomic(gmp,μ)
+opt = atomic(gmp, μ, 1e-03)
 
 # This time the atom extraction succeeds, which proves optimality of the moment relaxation.