+++ examples/volume.jl
@@ -39,7 +39,7 @@ using Plots
 
 
 # relaxation order for the Generalized Moment Problem
-order = 6
+order = 8 
 
 # polnomial variables
 @polyvar x y
@@ -65,20 +65,14 @@ leb_mom(i,j) = ((1-(-1)^(i+1))/(i+1))*((1-(-1)^(j+1))/(j+1))/4
 # DynamicPolynomials.jl provides the possibility to define a monomial vector.
 mons = monomials([x,y],0:2*order)
 # The monomial vector is not of a polynomial type, we need to convert it first.
-pons = convert(Vector{Polynomial{true,Int64}},mons)
-# The fiels mons.Z provides the exponents of the monomials of mons (and pons)
-cons = MomCons( Mom.(pons,μ)+Mom.(pons,ν),:eq,[leb_mom(mons.Z[i]...) for i = 1:length(mons.Z)])  
+pons = polynomial.(mons)
 
-# The constraints are added to the gmp model. Note that we gave a name to the constraints first 
+# The fiels mons.Z provides the exponents of the monomials of mons (and pons)
+exponents = mons.Z
+# The constraints are added to the gmp model. Note that we give a name to the constraints 
 # in order to be able to refer to them later on
 
-@mconstraints gmp cons
-
-#Note that we gave a name to the constraints in order to be able to refer to them later on.
-#we could have done 
-# @mconstraints gmp MomCons( Mom(pons,μ)+Mom(pons,ν),:eq,[leb_mom(mons.Z[i]...) for i = 1:length(mons.Z)])  
-# directly 
-
+@constraint gmp cons[i=1:length(exponents)] Mom(pons[i],μ)+Mom(pons[i],ν) == leb_mom(exponents[i]...) 
 
 # In order to relax the problem, we need to specify the relaxation order and a solver factory:
 relax!(gmp,order,with_optimizer(Mosek.Optimizer))