+++ test/runtests.jl
@@ -4,6 +4,7 @@ using DynamicPolynomials
 using SemialgebraicSets
 using SumOfSquares
 
+using CSDP
 
 @testset "MomentOpt Tests" begin
 
@@ -12,5 +13,6 @@ include("momexpr.jl")
 include("momcon.jl")
 include("macros.jl")
 include("show.jl")
+include("relax.jl")
 
 end