+++ test/show.jl
@@ -1,12 +1,23 @@
 @testset "Show Test" begin
     gmp = GMPModel()
-    @polyvar x y
     @test sprint(show, gmp) == """
-GMPModel:
-Feasability problem:
-s.t.
-
-Unknowns:
-OPTIMIZE_NOT_CALLED
-"""
-end
+    GMPModel:
+    Feasability problem:
+    s.t.
+    MomCon[]
+    Unknowns:
+    OPTIMIZE_NOT_CALLED
+    """
+    @polyvar x y
+    @measure gmp μ [x,y]
+    @measure gmp ν [x]
+    @constraint gmp Mom(x,μ)+Mom(x,ν) == 1
+    @test sprint(show, gmp) == """
+    GMPModel:
+    Feasability problem:
+    s.t.
+    MomCon[⟨μ, x⟩ + ⟨ν, x⟩ = 1.0]
+    Unknowns: μ ν
+    OPTIMIZE_NOT_CALLED
+    """
+end